Adding concept of CorrelationContext
Initial work to implement a correlation context and accessor which behave very much like the HttpContext.

IHttpContextAccessor is not registered by default and while it could be registered and used to get the value of TraceIdentifier it's a slightly heavy handed approach.

This work is a breaking change and will become 2.0.0 once released.